refactor(app): rename sessionCtrl to WebAppCtrl
