+++ client/app/index.html
@@ -7,7 +7,7 @@
         <title>OVH</title>
         <link rel="shortcut icon" href="../../assets/images/favicon.ico">
     </head>
-    <body data-ng-app="App" data-ng-controller="SessionCtrl" data-ng-cloak class="ng-cloak">
+    <body data-ng-app="App" data-ng-controller="WebAppCtrl as $ctrl" data-ng-cloak class="ng-cloak">
         <div class="col-container">
 
             <!-- Skip to main content -->
